zapisy/requirements.common.txt
Django~=3.0.4
django-bootstrap-pagination==1.7.0
django-cas-ng==4.1.1
django-crispy-forms==1.9.0
choicesenum==0.5.3
django-environ==0.4.4
django-extensions==2.2.9
django-filter==2.2.0
django-mailer==1.0.0
redis==3.4.1
rq==1.3.0
django-rq==2.3.0
django-test-without-migrations==0.6
django-webpack-loader==0.6.0
djangorestframework==3.11.0
django-pagedown~=2.0.3

Fabric3==1.14.post1
ipdb==0.11
prompt-toolkit==3.0.5
ipython==7.13.0
pycryptodome==3.8.1
rollbar==0.13.17
newrelic==2.68.0.50
gunicorn==19.7.1
python-dateutil==2.7.3
vobject==0.9.5
python-memcached==1.58
xhtml2pdf==0.2.2
# BBCode -> HTML
postmarkup==1.2.2
# Remove accents from strings (diacritics converted to Latin equivalents)
unidecode==1.0.22
# PyYAML is used for fixtures.
pyyaml==5.1
# Bokeh - library used for plotting in views displaying Poll results
bokeh==1.2.0

redis==2.10.6
gspread==3.1.0
typing-extensions==3.7.4.1
oauth2client==4.1.3
# We use html2text for one-off migration from HTML to markdown.
html2text